requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
minio
redis